+++ CHANGELOG.md
@@ -69,6 +69,7 @@ lists they are listed as multiple entries.
 - `scrollbar` theme option now also accepts `None` as a valid value to hide all scrollbars in rmpc
 - `TogglePause` in both the keybind and CLI to issue play if the current state is stopped
 - Made lyrics index matching more lenient
+- Changed the logging path from `/tmp/rmpc.log` to `/tmp/rmpc_${UID}.log`
 
 ### Fixed